trying to merge

+++ project-manager-workspace/ProjectManagerBasicBackend/src/main/java/com/ibm/projectManagerBasic/service/ServiceClass.java
@@ -18,24 +18,46 @@ public class ServiceClass {
 	
 	//add user data into the database
 	public boolean addNewUser(UserDetails user) {
-		boolean isPhoneNumUsed = false;
-		List<String> listOfPhoneNum = repo.getAllPhoneNum();
-		if(listOfPhoneNum.size() > 1) {
-			for(String phoneNum: listOfPhoneNum) {
-				if(phoneNum.equals(user.getUserPhoneNum())) {
-					isPhoneNumUsed = true;
-				}
-			}
-		}
-		if(!isPhoneNumUsed) {
+		boolean isEmailIdUsed = this.searchForEmailId(user.getUserEmailId());
+		if(!isEmailIdUsed) {
 			repo.save(user);
 		}
-		return isPhoneNumUsed;
+		return isEmailIdUsed;
+	}
+	
+	//Function to check entered email ID
+	public boolean searchForEmailId(String emailId) {
+		boolean isEmailIdUsed = false;
+		List<String> listOfEmailId = repo.getAllEmailId();
+		
+		for(String email: listOfEmailId) {
+			if(email.equals(emailId)) {
+				isEmailIdUsed = true;
+			}
+		}
+		return isEmailIdUsed;
+	}
+	
+	//Function for searching valid userId
+	public boolean searchForId(Integer id) {
+		boolean isIdFound = false;
+		List<Integer> listOfIds = repo.getAllId();
+		
+		for(int userId: listOfIds) {
+			if(id == userId) {
+				isIdFound = true;
+			}
+		}
+		return isIdFound;
 	}
 	
 	//edit user data by userId
-	public void updateUserDetails(UserDetails user, Integer id) {
-		repo.updateUserDetails(user.getUserName(), user.getUserPassword(), user.getUserDesignation(), id);
+	public boolean updateUserDetails(UserDetails user, Integer id) {
+		boolean isIdFound = this.searchForId(id);
+		if(isIdFound) {
+			repo.updateUserDetails(user.getUserName(), user.getUserEmailId(),user.getUserPassword(), user.getUserDesignation(), id);
+		}
+		return isIdFound;
 	}
 
 	//view all of the user information in database
@@ -44,8 +66,12 @@ public class ServiceClass {
 	}
 	
 	//delete user data by userId
-	public void deleteUserById(Integer id) {
-		repo.deleteById(id);
+	public boolean deleteUserById(Integer id) {
+		boolean isIdFound = this.searchForId(id);
+		if(isIdFound) {
+			repo.deleteById(id);
+		}
+		return isIdFound;
 	}
 	
 	//get users by id
@@ -57,12 +83,19 @@ public class ServiceClass {
 	public List<UserDetails> findUserByName(String userName) {
 		return repo.findByUserName(userName);
 	}
+	
+	//get user by emailID
+	public List<UserDetails> findByUserEmailId(String emailId){
+		return repo.findByUserEmailId(emailId);
+	}
 
 	//check the user password from database
 	public int checkPassword(UserDetails user) {
 		int condition = -1;
-		if(this.findUserByName(user.getUserName()).size() > 0) {
-			String password = repo.checkPassword(user.getUserName());
+		String userEmail = user.getUserEmailId();
+		boolean isEmailIdFound = this.searchForEmailId(userEmail);
+		if(isEmailIdFound) {
+			String password = repo.checkPassword(userEmail);
 		    boolean isPasswordMatched = password.equals(user.getUserPassword());
 		    if(!isPasswordMatched) {
 		    	condition = 0;
